dashboard/dashboard_requirements.txt
streamlit
pandas
altair
pymssql
python-dotenv
pyathena